refactor: move invalid duration embed to general
